+++ bin/.gitignore
@@ -1,1 +1,2 @@
 /agentes/
+/restaurante/
